+++ guidata/dataset/__init__.py
@@ -47,6 +47,7 @@ from .datatypes import (
     EndTabGroup,
     FormatProp,
     FuncProp,
+    FuncPropMulti,
     GetAttrProp,
     GroupItem,
     ItemProperty,