+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 💥 New features:
 
+* New operator property `FuncPropMulti` for handling multiple properties:
+  * This property allows you to apply a function to multiple item properties at once.
+  * It can be used to create more complex dependencies between items in a dataset.
+  * See the `guidata.tests.dataset.test_activable_items` module for an example of usage.
+
 * New script `gbuild` for building the package:
   * This script is a wrapper around the `guidata.utils.securebuild` module, which ensures that the build process is secure and reproducible.
   * It checks that the `pyproject.toml` file is present in the root of the repository, and that it is committed to Git.